function into file, sugestion
